Move E2EExpr unit tests back into E2EExpr.py
