+++ sycl/test/e2e_test_requirements/check_e2eexpr_logic.cpp
@@ -3,6 +3,6 @@
 // statements. This test runs the unit tests related to these expressions.
 //
 // REQUIRES: linux
-// DEFINE: %{e2e_folder}=%S/../../test-e2e
+// DEFINE: %{E2EExpr}=%S/../../test-e2e/E2EExpr.py
 // DEFINE: %{lit_source}=%S/../../../llvm/utils/lit
-// RUN: env PYTHONPATH=%{lit_source}:%{e2e_folder} python %S/E2EExpr_tests.py
+// RUN: env PYTHONPATH=%{lit_source} python %{E2EExpr}
